Fix integer cast test initializers

+++ test/casts.test.ts
@@ -7,7 +7,7 @@ test("integer casts execute", async () => {
     fn main() -> i32 {
         let mut score: i32 = 0;
 
-        let neg_i8: i8 = -1;
+        let neg_i8: i8 = (-1 as i8);
         let neg_as_u8: u8 = neg_i8 as u8;
         if neg_as_u8 as i32 == 255 {
             score = score + 1;
@@ -16,7 +16,7 @@ test("integer casts execute", async () => {
             0
         };
 
-        let byte_value: u8 = 255;
+        let byte_value: u8 = (255 as u8);
         let byte_as_i16: i16 = byte_value as i16;
         if byte_as_i16 as i32 == 255 {
             score = score + 1;
@@ -25,7 +25,7 @@ test("integer casts execute", async () => {
             0
         };
 
-        let neg_i16: i16 = -300;
+        let neg_i16: i16 = (-300 as i16);
         let neg_as_u16: u16 = neg_i16 as u16;
         if neg_as_u16 as i32 == 65236 {
             score = score + 1;
@@ -51,7 +51,7 @@ test("integer casts execute", async () => {
             0
         };
 
-        let small_u8: u8 = 128;
+        let small_u8: u8 = (128 as u8);
         let small_as_i64: i64 = small_u8 as i64;
         if small_as_i64 as i32 == 128 {
             score = score + 1;
@@ -60,7 +60,7 @@ test("integer casts execute", async () => {
             0
         };
 
-        let neg_small_i16: i16 = -1234;
+        let neg_small_i16: i16 = (-1234 as i16);
         let neg_small_i64: i64 = neg_small_i16 as i64;
         if neg_small_i64 as i32 == -1234 {
             score = score + 1;
